Decisions-focused learning paper

Public

DecisionFocusedLearningBenchmarks.FixedSizeShortestPath.FixedSizeShortestPathBenchmark — Type

struct FixedSizeShortestPathBenchmark <: AbstractBenchmark

Benchmark problem for the shortest path problem. In this benchmark, all graphs are acyclic directed grids, all of the same size grid_size. Features are given at instance level (one dimensional vector of length p for each graph).

Data is generated using the process described in: https://arxiv.org/abs/2307.13565.

Fields

graph::Graphs.SimpleGraphs.SimpleDiGraph{Int64}: grid graph instance

grid_size::Tuple{Int64, Int64}: grid size of graphs

p::Int64: size of feature vectors

deg::Int64: degree of formula between features and true weights

ν::Float32: multiplicative noise for true weights sampled between [1-ν, 1+ν], should be between 0 and 1

DecisionFocusedLearningBenchmarks.FixedSizeShortestPath.FixedSizeShortestPathBenchmark — Method

FixedSizeShortestPathBenchmark(
;
    grid_size,
    p,
    deg,
    ν
) -> FixedSizeShortestPathBenchmark

Constructor for FixedSizeShortestPathBenchmark.

DecisionFocusedLearningBenchmarks.PortfolioOptimization.PortfolioOptimizationBenchmark — Type

struct PortfolioOptimizationBenchmark <: AbstractBenchmark

Benchmark problem for the portfolio optimization problem.

Data is generated using the process described in: https://arxiv.org/abs/2307.13565.

Fields

d::Int64: number of assets

p::Int64: size of feature vectors

deg::Int64: hypermarameter for data generation

ν::Float32: another hyperparameter, should be positive

Σ::Matrix{Float32}: covariance matrix

γ::Float32: maximum variance of portfolio

L::Matrix{Float32}: useful for dataset generation

f::Vector{Float32}: useful for dataset generation

DecisionFocusedLearningBenchmarks.PortfolioOptimization.PortfolioOptimizationBenchmark — Method

PortfolioOptimizationBenchmark(
;
    d,
    p,
    deg,
    ν,
    seed
) -> PortfolioOptimizationBenchmark

Constructor for PortfolioOptimizationBenchmark.

DecisionFocusedLearningBenchmarks.SubsetSelection.SubsetSelectionBenchmark — Type

struct SubsetSelectionBenchmark <: AbstractBenchmark

Benchmark problem for the subset selection problem. Reference: https://arxiv.org/abs/2307.13565.

The goal is to select the best k items from a set of n items, without knowing their values, but only observing some features.

Fields

n::Int64: total number of items

k::Int64: number of items to select

Private

DecisionFocusedLearningBenchmarks.Utils.generate_dataset — Function

generate_dataset(
    bench::FixedSizeShortestPathBenchmark;
    ...
) -> Vector
generate_dataset(
    bench::FixedSizeShortestPathBenchmark,
    dataset_size::Int64;
    seed,
    type
) -> Vector

Generate dataset for the shortest path problem.

DecisionFocusedLearningBenchmarks.Utils.generate_maximizer — Method

generate_maximizer(
    bench::FixedSizeShortestPathBenchmark;
    use_dijkstra
) -> DecisionFocusedLearningBenchmarks.FixedSizeShortestPath.var"#shortest_path_maximizer#8"{DecisionFocusedLearningBenchmarks.FixedSizeShortestPath.var"#shortest_path_maximizer#5#9"{typeof(Graphs.dijkstra_shortest_paths), Vector{Int64}, Vector{Int64}, Int64, Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}}}

Outputs a function that computes the longest path on the grid graph, given edge weights θ as input.

maximizer = generate_maximizer(bench)
maximizer(θ)

DecisionFocusedLearningBenchmarks.Utils.generate_statistical_model — Method

generate_statistical_model(
    bench::FixedSizeShortestPathBenchmark
) -> Flux.Chain{T} where T<:Tuple{Flux.Dense{typeof(identity), Matrix{Float32}}}

Initialize a linear model for bench using Flux.

DecisionFocusedLearningBenchmarks.Utils.generate_dataset — Function

generate_dataset(
    bench::PortfolioOptimizationBenchmark;
    ...
) -> Vector
generate_dataset(
    bench::PortfolioOptimizationBenchmark,
    dataset_size::Int64;
    seed,
    type
) -> Vector

Generate a dataset of labeled instances for the portfolio optimization problem.

DecisionFocusedLearningBenchmarks.Utils.generate_maximizer — Method

generate_maximizer(
    bench::PortfolioOptimizationBenchmark
) -> DecisionFocusedLearningBenchmarks.PortfolioOptimization.var"#portfolio_maximizer#3"{Float32, Matrix{Float32}, Int64}

Create a function solving the MIQP formulation of the portfolio optimization problem.

DecisionFocusedLearningBenchmarks.Utils.generate_statistical_model — Method

generate_statistical_model(
    bench::PortfolioOptimizationBenchmark
) -> Flux.Dense{typeof(identity), Matrix{Float32}}

Initialize a linear model for bench using Flux.

DecisionFocusedLearningBenchmarks.Utils.generate_dataset — Function

generate_dataset(
    bench::SubsetSelectionBenchmark;
    ...
) -> Any
generate_dataset(
    bench::SubsetSelectionBenchmark,
    dataset_size::Int64;
    seed,
    identity_mapping
) -> Any

Generate a dataset of labeled instances for the subset selection problem. The mapping between features and cost is identity.

DecisionFocusedLearningBenchmarks.Utils.generate_maximizer — Method

generate_maximizer(
    bench::SubsetSelectionBenchmark
) -> Base.Fix2{typeof(DecisionFocusedLearningBenchmarks.SubsetSelection.top_k), Int64}

Return a top k maximizer.

DecisionFocusedLearningBenchmarks.Utils.generate_statistical_model — Method

generate_statistical_model(
    bench::SubsetSelectionBenchmark;
    seed
) -> Flux.Dense{typeof(identity), Matrix{Float32}}

Initialize a linear model for bench using Flux.