Dynamic Vehicle Scheduling

Public

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DynamicVehicleSchedulingBenchmark — Type

struct DynamicVehicleSchedulingBenchmark <: AbstractDynamicBenchmark{true}

Abstract type for dynamic vehicle scheduling benchmarks.

Fields

max_requests_per_epoch::Int64: maximum number of customers entering the system per epoch

Δ_dispatch::Float64: time between decision and dispatch of a vehicle

epoch_duration::Float64: duration of an epoch

two_dimensional_features::Bool: whether to use two-dimensional features

Private

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPEnv — Method

DVSPEnv(
    instance::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.Instance;
    seed
) -> DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPEnv{DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPState{DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.StaticInstance{Float64}}, Random.MersenneTwister, Nothing}

Constructor for DVSPEnv.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPState — Type

State data structure for the Dynamic Vehicle Scheduling Problem.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.Instance — Type

struct Instance{I<:DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.StaticInstance, T}

Instance data structure for the dynamic vehicle scheduling problem.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.KleopatraVSPPolicy — Type

struct KleopatraVSPPolicy{P}

Kleopatra policy for the Dynamic Vehicle Scheduling Problem.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.Point — Type

struct Point{T}

Basic point structure.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.StaticInstance — Type

struct StaticInstance{T}

Instance data structure for the (deterministic and static) Vehicle Scheduling Problem.

Fields

coordinate::Array{DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.Point{T}, 1} where T: coordinates of the locations. The first one is always the depot.

service_time::Vector: service time at each location

start_time::Vector: start time at each location

duration::Matrix: duration matrix between locations

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.VSPSolution — Type

struct VSPSolution

Solution for the static Vehicle Scheduling Problem.

Fields

routes::Vector{Vector{Int64}}: list of routes, each route being a list of request indices in corresponding instance (excluding the depot).

edge_matrix::BitMatrix: size (nblocations, nblocations). edge_matrix[i, j] is equal to 1 if a route takes edge (i, j).

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.VSPSolution — Method

VSPSolution(
    routes::Vector{Vector{Int64}};
    max_index
) -> DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.VSPSolution

Build a VSPSolution from routes. Set max_index to manually define the size of the edge_index matrix.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.anticipative_solver — Function

anticipative_solver(
    env::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPEnv;
    ...
) -> Tuple{Union{Float64, Vector{Float64}}, Vector}
anticipative_solver(
    env::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPEnv,
    scenario;
    model_builder,
    two_dimensional_features,
    reset_env,
    nb_epochs,
    seed
) -> Tuple{Union{Float64, Vector{Float64}}, Vector}

Solve the anticipative VSP problem for environment env. For this, it uses the current environment history, so make sure that the environment is terminated before calling this method.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.apply_routes! — Method

remove dispatched customers, and update must-dispatch and postponable flags.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.coordinate — Method

coordinate(
    state::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPState
) -> Array{DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.Point{T}, 1} where T

Get the coordinates vector.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.coordinate — Method

coordinate(
    instance::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.StaticInstance
) -> Array{DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.Point{T}, 1} where T

Get the coordinates vector.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.cost — Method

cost(
    routes::Vector{Vector{Int64}},
    duration::AbstractMatrix
) -> Any

Compute the total cost of a set of routes given a distance matrix, i.e. the sum of the distances between each location in the route. Note that the first location is implicitly assumed to be the depot, and should not appear in the route.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.create_graph — Method

create_graph(
    state::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPState
) -> Graphs.SimpleGraphs.SimpleDiGraph{Int64}

Create the acyclic digraph associated with the given VSP state.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.create_graph — Method

create_graph(
    instance::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.StaticInstance
) -> Graphs.SimpleGraphs.SimpleDiGraph{Int64}

Create the acyclic digraph associated with the given VSP instance.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.customer_count — Method

customer_count(
    state::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPState
) -> Int64

Return the number of customers in state.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.customer_count — Method

customer_count(
    instance::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.StaticInstance
) -> Int64

Return the number of customers in instance (excluding the depot).

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.duration — Method

duration(
    state::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPState
) -> Matrix

Get the duration matrix.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.duration — Method

duration(
    instance::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.StaticInstance
) -> Matrix

Get the duration matrix.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.edge_matrix — Method

edge_matrix(
    solution::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.VSPSolution
) -> BitMatrix

Get edge matrix from solution.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.is_feasible — Method

is_feasible(
    state::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPState,
    routes::Vector{Vector{Int64}};
    verbose
) -> Bool

Check if the given routes are feasible. Routes should be given with global indexation. Use env_routes_from_state_routes if needed to convert the indices beforehand.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.location_count — Method

location_count(
    state::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPState
) -> Int64

Return the number of locations in state (customers + depot).

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.location_count — Method

location_count(
    instance::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.StaticInstance
) -> Int64

Return the number of locations in instance (customers + depot).

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.planning_start_time — Method

planning_start_time(
    env::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPEnv
) -> Any

Get the planning start time of the environment, i.e. the time at which vehicles routes dispatched in current epoch can depart.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.plot_instance — Method

plot_instance(
    instance::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.StaticInstance;
    customer_markersize,
    depot_markersize,
    alpha_depot,
    customer_color,
    depot_color,
    kwargs...
) -> Plots.Plot

Plot the given static VSP instance.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.prize_collecting_vsp — Method

prize_collecting_vsp(
    θ::AbstractVector;
    instance,
    model_builder,
    kwargs...
)

Solve the Prize Collecting Vehicle Scheduling Problem defined by instance and prize vector θ.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.read_vsp_instance — Method

read_vsp_instance(
    filepath::String;
    rounded,
    normalization
) -> DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.StaticInstance

Create a VSPInstance from file filepath containing a VRPTW instance. It uses time window values to compute task times as the middle of the interval.

Round all values to Int if rounded=true. Normalize all time values by the normalization parameter.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.retrieve_routes — Method

retrieve_routes(
    y::AbstractArray,
    graph::Graphs.AbstractGraph
) -> Vector{Vector{Int64}}

Retrieve routes solution from the given MIP solution y matrix and graph.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.retrieve_routes_anticipative — Method

retrieve_routes_anticipative(
    y::AbstractArray,
    dvspenv::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPEnv,
    customer_index,
    epoch_indices
) -> Any

Retrieve anticipative routes solution from the given MIP solution y. Outputs a set of routes per epoch.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.routes — Method

routes(
    solution::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.VSPSolution
) -> Vector{Vector{Int64}}

Get routes from solution.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.sample_indices — Method

sample_indices(rng::Random.AbstractRNG, k, N) -> Any

Sample k random different indices from 2 to N+1.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.service_time — Method

service_time(
    state::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPState
) -> Vector

Get the service time vector

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.service_time — Method

service_time(
    instance::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.StaticInstance
) -> Vector

Get the service time vector.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.start_time — Method

start_time(
    state::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPState
) -> Vector

Get the start time vector.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.start_time — Method

start_time(
    instance::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.StaticInstance
) -> Vector

Get the start time vector.

DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.time — Method

time(
    env::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPEnv
) -> Any

Get the current time of the environment, i.e. the start time of the current_epoch.

DecisionFocusedLearningBenchmarks.Utils.is_terminated — Method

is_terminated(
    env::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPEnv
) -> Bool

Check if the episode is terminated, i.e. if the current epoch is the last one.

DecisionFocusedLearningBenchmarks.Utils.observe — Method

observe(
    env::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPEnv
) -> Tuple{Any, DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPState}

Get the current state of the environment.

DecisionFocusedLearningBenchmarks.Utils.reset! — Method

reset!(
    env::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPEnv;
    seed,
    reset_rng
)

Reset the environment to its initial state. Also reset the rng to seed if reset_rng is set to true.

DecisionFocusedLearningBenchmarks.Utils.step! — Function

step!(
    env::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPEnv,
    routes
) -> Any
step!(
    env::DecisionFocusedLearningBenchmarks.DynamicVehicleScheduling.DVSPEnv,
    routes,
    scenario
) -> Any

Remove dispatched customers, advance time, and add new requests to the environment.