Interface

Public

InferOptBenchmarks.Utils.AbstractBenchmark — Type

abstract type AbstractBenchmark

Abstract type interface for a benchmark problem.

The following methods exist for benchmarks:

generate_dataset

generate_statistical_model

generate_maximizer

plot_data

objective_value

compute_gap

InferOptBenchmarks.Utils.DataSample — Type

struct DataSample{F, S, C, I}

Data sample data structure.

Fields

x::Any: features

θ::Any: costs

y::Any: solution

instance::Any: instance

InferOptBenchmarks.Utils.average_tensor — Method

average_tensor(x)

Average the tensor x along its third axis.

InferOptBenchmarks.Utils.compute_gap — Function

compute_gap(::AbstractBenchmark, dataset::Vector{<:DataSample}, statistical_model, maximizer) -> Float64

Compute the average relative optimality gap of the pipeline on the dataset.

InferOptBenchmarks.Utils.compute_gap — Method

compute_gap(
    bench::AbstractBenchmark,
    dataset::Vector{<:DataSample},
    statistical_model,
    maximizer
) -> Any

Default behaviour of compute_gap for a benchmark problem where features, solutions and costs are all defined.

InferOptBenchmarks.Utils.generate_dataset — Function

generate_dataset(::AbstractBenchmark, dataset_size::Int) -> Vector{<:DataSample}

Generate a Vector of DataSample of length dataset_size for given benchmark. Content of the dataset can be visualized using plot_data, when it applies.

InferOptBenchmarks.Utils.generate_maximizer — Function

generate_maximizer(::AbstractBenchmark)

Generates a maximizer function. Returns a callable f: (θ; kwargs...) -> y, where θ is a cost array and y is a solution.

InferOptBenchmarks.Utils.generate_statistical_model — Function

generate_statistical_model(::AbstractBenchmark)

Initializes and return an untrained statistical model of the CO-ML pipeline. It's usually a Flux model, that takes a feature matrix x as iinput, and returns a cost array θ as output.

InferOptBenchmarks.Utils.get_path — Method

get_path(
    parents::AbstractVector{<:Integer},
    s::Integer,
    d::Integer
) -> Vector{T} where T<:Integer

Retrieve a path from the parents array and start sand endd`` of path.

InferOptBenchmarks.Utils.grid_graph — Method

grid_graph(
    costs::AbstractArray{R, 2};
    acyclic
) -> SimpleWeightedGraphs.SimpleWeightedDiGraph{Int64}

Convert a grid of cell costs into a weighted directed graph from SimpleWeightedGraphs.jl, where the vertices correspond to the cells and the edges are weighted by the cost of the arrival cell.

If acyclic = false, a cell has edges to each one of its 8 neighbors.

If acyclic = true, a cell has edges to its south, east and southeast neighbors only (ensures an acyclic graph where topological sort will work)

This can be used to model the Warcraft shortest paths problem of

Differentiation of Blackbox Combinatorial Solvers, Vlastelica et al. (2019)

InferOptBenchmarks.Utils.neg_tensor — Method

neg_tensor(x)

Compute minus softplus element-wise on tensor x.

InferOptBenchmarks.Utils.path_to_matrix — Method

path_to_matrix(
    path::Vector{<:Integer},
    h::Integer,
    w::Integer
) -> Matrix{Int64}

Transform path into a binary matrix of size (h, w) where each cell is 1 if the cell is part of the path, 0 otherwise.

InferOptBenchmarks.Utils.plot_data — Function

plot_data(::AbstractBenchmark, args...)

Plot a data sample from the dataset created by generate_dataset. Check the specific benchmark documentation of plot_data for more details on the arguments.

InferOptBenchmarks.Utils.squeeze_last_dims — Method

squeeze_last_dims(x)

Squeeze two last dimensions on tensor x.

Private

InferOptBenchmarks.Utils.coord_to_index — Method

coord_to_index(
    i::Integer,
    j::Integer,
    h::Integer,
    w::Integer
) -> Any

Given a pair of row-column coordinates (i, j) on a grid of size (h, w), compute the corresponding vertex index in the graph generated by grid_graph.

InferOptBenchmarks.Utils.count_edges — Method

count_edges(h::Integer, w::Integer; acyclic)

Compute the number of edges in a grid graph of size (h, w).

InferOptBenchmarks.Utils.index_to_coord — Method

index_to_coord(
    v::Integer,
    h::Integer,
    w::Integer
) -> Tuple{Any, Any}

Given a vertex index in the graph generated by grid_graph, compute the corresponding row-column coordinates (i, j) on a grid of size (h, w).

InferOptBenchmarks.Utils.objective_value — Method

objective_value(
    _::AbstractBenchmark,
    θ̄::AbstractArray,
    y::AbstractArray
) -> Any

Default behaviour of objective_value.