Subset selection

Public

DecisionFocusedLearningBenchmarks.SubsetSelection.SubsetSelectionBenchmark — Type

struct SubsetSelectionBenchmark <: AbstractBenchmark

Benchmark problem for the subset selection problem. Reference: https://arxiv.org/abs/2307.13565.

The goal is to select the best k items from a set of n items, without knowing their values, but only observing some features.

Fields

n::Int64: total number of items

k::Int64: number of items to select

DecisionFocusedLearningBenchmarks.Utils.generate_dataset — Function

generate_dataset(
    bench::SubsetSelectionBenchmark;
    ...
) -> Any
generate_dataset(
    bench::SubsetSelectionBenchmark,
    dataset_size::Int64;
    seed,
    identity_mapping
) -> Any

Generate a dataset of labeled instances for the subset selection problem. The mapping between features and cost is identity.

DecisionFocusedLearningBenchmarks.Utils.generate_maximizer — Method

generate_maximizer(
    bench::SubsetSelectionBenchmark
) -> Base.Fix2{typeof(DecisionFocusedLearningBenchmarks.SubsetSelection.top_k), Int64}

Return a top k maximizer.

DecisionFocusedLearningBenchmarks.Utils.generate_statistical_model — Method

generate_statistical_model(
    bench::SubsetSelectionBenchmark;
    seed
) -> Flux.Dense{typeof(identity), Matrix{Float32}}

Initialize a linear model for bench using Flux.

Private