Warcraft

Public

InferOptBenchmarks.Warcraft.coord_to_index — Method

coord_to_index(i, j, h, w)

Given a pair of row-column coordinates (i, j) on a grid of size (h, w), compute the corresponding vertex index in the graph generated by warcraft_grid_graph.

InferOptBenchmarks.Warcraft.index_to_coord — Method

index_to_coord(v, h, w)

Given a vertex index in the graph generated by warcraft_grid_graph, computethe corresponding row-column coordinates (i, j) on a grid of size (h, w).

InferOptBenchmarks.Warcraft.warcraft_grid_graph — Method

warcraft_grid_graph(costs::AbstractMatrix; acyclic::Bool=false)

Convert a grid of Warcraft cell costs into a weighted directed graph from SimpleWeightedGraphs.jl, where the vertices correspond to the cells and the edges are weighted by the cost of the arrival cell.

This represents the Warcraft shortest paths problem of

Differentiation of Blackbox Combinatorial Solvers, Vlastelica et al. (2019)

If acyclic = false, a cell has edges to each one of its 8 neighbors.

If acyclic = true, a cell has edges to its south, east and southeast neighbors only (ensures an acyclic graph where topological sort will work)

Private